move approx_percentile_cont_accumulator to function-aggregate
