+++ datafusion/functions-aggregate/src/approx_percentile_cont.rs
@@ -1,0 +1,249 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+
+use arrow::{
+    array::{
+        ArrayRef, Float32Array, Float64Array, Int16Array, Int32Array, Int64Array,
+        Int8Array, UInt16Array, UInt32Array, UInt64Array, UInt8Array,
+    },
+    datatypes::DataType,
+};
+
+use datafusion_common::{downcast_value, internal_err, ScalarValue};
+use datafusion_expr::Accumulator;
+use datafusion_physical_expr_common::aggregate::tdigest::{DEFAULT_MAX_SIZE, TDigest, TryIntoF64};
+
+#[derive(Debug)]
+pub struct ApproxPercentileAccumulator {
+    digest: TDigest,
+    percentile: f64,
+    return_type: DataType,
+}
+
+impl ApproxPercentileAccumulator {
+    pub fn new(percentile: f64, return_type: DataType) -> Self {
+        Self {
+            digest: TDigest::new(DEFAULT_MAX_SIZE),
+            percentile,
+            return_type,
+        }
+    }
+
+    pub fn new_with_max_size(
+        percentile: f64,
+        return_type: DataType,
+        max_size: usize,
+    ) -> Self {
+        Self {
+            digest: TDigest::new(max_size),
+            percentile,
+            return_type,
+        }
+    }
+
+    pub(crate) fn merge_digests(&mut self, digests: &[TDigest]) {
+        let digests = digests.iter().chain(std::iter::once(&self.digest));
+        self.digest = TDigest::merge_digests(digests)
+    }
+
+    pub(crate) fn convert_to_float(values: &ArrayRef) -> datafusion_common::Result<Vec<f64>> {
+        match values.data_type() {
+            DataType::Float64 => {
+                let array = downcast_value!(values, Float64Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::Float32 => {
+                let array = downcast_value!(values, Float32Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::Int64 => {
+                let array = downcast_value!(values, Int64Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::Int32 => {
+                let array = downcast_value!(values, Int32Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::Int16 => {
+                let array = downcast_value!(values, Int16Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::Int8 => {
+                let array = downcast_value!(values, Int8Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::UInt64 => {
+                let array = downcast_value!(values, UInt64Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::UInt32 => {
+                let array = downcast_value!(values, UInt32Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::UInt16 => {
+                let array = downcast_value!(values, UInt16Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            DataType::UInt8 => {
+                let array = downcast_value!(values, UInt8Array);
+                Ok(array
+                    .values()
+                    .iter()
+                    .filter_map(|v| v.try_as_f64().transpose())
+                    .collect::<datafusion_common::Result<Vec<_>>>()?)
+            }
+            e => internal_err!(
+                "APPROX_PERCENTILE_CONT is not expected to receive the type {e:?}"
+            ),
+        }
+    }
+}
+
+impl Accumulator for ApproxPercentileAccumulator {
+    fn state(&mut self) -> datafusion_common::Result<Vec<ScalarValue>> {
+        Ok(self.digest.to_scalar_state().into_iter().collect())
+    }
+
+    fn update_batch(&mut self, values: &[ArrayRef]) -> datafusion_common::Result<()> {
+        let values = &values[0];
+        let sorted_values = &arrow::compute::sort(values, None)?;
+        let sorted_values = ApproxPercentileAccumulator::convert_to_float(sorted_values)?;
+        self.digest = self.digest.merge_sorted_f64(&sorted_values);
+        Ok(())
+    }
+
+    fn evaluate(&mut self) -> datafusion_common::Result<ScalarValue> {
+        if self.digest.count() == 0.0 {
+            return ScalarValue::try_from(self.return_type.clone());
+        }
+        let q = self.digest.estimate_quantile(self.percentile);
+
+        // These acceptable return types MUST match the validation in
+        // ApproxPercentile::create_accumulator.
+        Ok(match &self.return_type {
+            DataType::Int8 => ScalarValue::Int8(Some(q as i8)),
+            DataType::Int16 => ScalarValue::Int16(Some(q as i16)),
+            DataType::Int32 => ScalarValue::Int32(Some(q as i32)),
+            DataType::Int64 => ScalarValue::Int64(Some(q as i64)),
+            DataType::UInt8 => ScalarValue::UInt8(Some(q as u8)),
+            DataType::UInt16 => ScalarValue::UInt16(Some(q as u16)),
+            DataType::UInt32 => ScalarValue::UInt32(Some(q as u32)),
+            DataType::UInt64 => ScalarValue::UInt64(Some(q as u64)),
+            DataType::Float32 => ScalarValue::Float32(Some(q as f32)),
+            DataType::Float64 => ScalarValue::Float64(Some(q)),
+            v => unreachable!("unexpected return type {:?}", v),
+        })
+    }
+
+    fn merge_batch(&mut self, states: &[ArrayRef]) -> datafusion_common::Result<()> {
+        if states.is_empty() {
+            return Ok(());
+        }
+
+        let states = (0..states[0].len())
+            .map(|index| {
+                states
+                    .iter()
+                    .map(|array| ScalarValue::try_from_array(array, index))
+                    .collect::<datafusion_common::Result<Vec<_>>>()
+                    .map(|state| TDigest::from_scalar_state(&state))
+            })
+            .collect::<datafusion_common::Result<Vec<_>>>()?;
+
+        self.merge_digests(&states);
+
+        Ok(())
+    }
+
+    fn size(&self) -> usize {
+        std::mem::size_of_val(self) + self.digest.size()
+            - std::mem::size_of_val(&self.digest)
+            + self.return_type.size()
+            - std::mem::size_of_val(&self.return_type)
+    }
+}
+
+
+#[cfg(test)]
+mod tests {
+    use arrow_schema::DataType;
+
+    use datafusion_physical_expr_common::aggregate::tdigest::TDigest;
+
+    use crate::approx_percentile_cont::ApproxPercentileAccumulator;
+
+    #[test]
+    fn test_combine_approx_percentile_accumulator() {
+        let mut digests: Vec<TDigest> = Vec::new();
+
+        // one TDigest with 50_000 values from 1 to 1_000
+        for _ in 1..=50 {
+            let t = TDigest::new(100);
+            let values: Vec<_> = (1..=1_000).map(f64::from).collect();
+            let t = t.merge_unsorted_f64(values);
+            digests.push(t)
+        }
+
+        let t1 = TDigest::merge_digests(&digests);
+        let t2 = TDigest::merge_digests(&digests);
+
+        let mut accumulator =
+            ApproxPercentileAccumulator::new_with_max_size(0.5, DataType::Float64, 100);
+
+        accumulator.merge_digests(&[t1]);
+        assert_eq!(accumulator.digest.count(), 50_000.0);
+        accumulator.merge_digests(&[t2]);
+        assert_eq!(accumulator.digest.count(), 100_000.0);
+    }
+}
