+++ datafusion/functions-aggregate/src/lib.rs
@@ -61,6 +61,7 @@ pub mod median;
 pub mod stddev;
 pub mod sum;
 pub mod variance;
+pub mod approx_percentile_cont;
 
 use datafusion_common::Result;
 use datafusion_execution::FunctionRegistry;
